backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0
pandas==2.2.2
openpyxl==3.1.5
google-genai>=1.0.0
python-dotenv==1.0.1
aiofiles==24.1.0
